Feat : 최근 검색어 컴포넌트 분리, 자동저장 기능 토글 구현
